[추가] 노티스 구현 (ViewModel 서버 연결 후 작업, 네비게이터 작업, 프로필 평가, auth)
