+++ UJeongApp/SceneDelegate.swift
@@ -22,3 +22,4 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     }
 }
 
+
